- Runner: Minor bug fix

+++ runner/internal/models/backend.go
@@ -1,12 +1,12 @@
 package models
 
 type Resource struct {
-	CPUs          int   `yaml:"cpus,omitempty"`
-	Memory        int   `yaml:"memory_mib,omitempty"`
-	GPUs          []GPU `yaml:"gpus,omitempty"`
-	Interruptible bool  `yaml:"interruptible,omitempty"`
-	ShmSize       int64 `yaml:"shm_size_mib,omitempty"`
-	Local         bool  `json:"local"`
+	CPUs          int    `yaml:"cpus,omitempty"`
+	Memory        uint64 `yaml:"memory_mib,omitempty"`
+	GPUs          []GPU  `yaml:"gpus,omitempty"`
+	Interruptible bool   `yaml:"interruptible,omitempty"`
+	ShmSize       int64  `yaml:"shm_size_mib,omitempty"`
+	Local         bool   `json:"local"`
 }
 
 type Job struct {
